requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
banal==1.0.6
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
cryptography==40.0.1
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
ecdsa==0.18.0
executing==1.2.0
fastapi==0.95.0
fastjsonschema==2.16.3
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
langchain==0.0.134
line-bot-sdk==2.4.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
redbird==0.7.1
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rocketry==2.5.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.41
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.1
widgetsnbextension==4.0.7
yarl==1.8.2
